Errorprone: NarrowingCompoundAssignment
Compound assignments from char to byte hide lossy casts
Compound assignments from int to byte hide lossy casts
Compound assignments from float to long hide lossy casts

Test: m Bluetooth
Bug: 344658662
Flag: Exempt, enforce error in the futur

+++ android/app/src/com/android/bluetooth/le_scan/AppScanStats.java
@@ -957,7 +957,7 @@ public class AppScanStats {
         int lowPowerScan = mLowPowerScan;
         int balancedScan = mBalancedScan;
         int lowLatencyScan = mLowLantencyScan;
-        int ambientDiscoveryScan = mAmbientDiscoveryScan;
+        long ambientDiscoveryScan = mAmbientDiscoveryScan;
 
         if (!mOngoingScans.isEmpty()) {
             for (Integer key : mOngoingScans.keySet()) {
